fix: emit JSON logs so the dashboard queries return rows
The saved Logs Insights queries and the dashboard's p95 widget
(terraform/modules/stack/alerting.tf) read log_processed.fields.* — the
shape Fluent Bit produces when it parses a JSON log line and nests it
under Merge_Log_Key. The app emitted human-readable text, so every one of
those queries returned zero rows against correct infrastructure.

Two changes, both required:

1. tracing_subscriber's JSON formatter (and its "json" feature). NOT
   flatten_event: the queries address fields as log_processed.fields.*,
   which is the formatter's default nesting.

2. latency_ms cast to u64. Duration::as_millis returns u128, and tracing
   has no Value impl for u128, so it fell back to Debug recording and
   serialized the STRING "5" rather than the number 5. CloudWatch can't
   avg()/pct() a string, so the p95 widget would have stayed empty with
   the formatter alone. status_code was already numeric (u16).

The producing event in middleware.rs already carried the right field
names; only the encoding was wrong.

Verified against real output, not inference: ran the binary on an
isolated Redis, wrapped each emitted line under log_processed to mimic
Fluent Bit, and evaluated the saved queries' field paths — latency_ms and
status_code both numeric, the 500 path matched by errors-recent. A Redis
outage shows up as a ~30s latency_ms (bb8's connection_timeout), which is
now visible on the p95 row rather than invisible.

+++ src/middleware.rs
@@ -155,9 +155,17 @@ pub fn create_trace_layer() -> TraceLayer<
                     span.record("http.response.body.size", content_length);
                 }
 
+                // `as u64` is load-bearing, not a lint fix. Duration::as_millis
+                // returns u128, and `tracing` has no Value impl for u128 — it
+                // falls back to Debug recording, which serializes as the STRING
+                // "5" instead of the number 5. CloudWatch Logs Insights can't
+                // avg()/pct() a string, so the dashboard's p95 widget and the
+                // p95-by-service saved query (terraform/modules/stack/
+                // alerting.tf) would still return empty with JSON logs on.
+                // u64 milliseconds overflows after ~584 million years.
                 tracing::info!(
                     status_code = response.status().as_u16(),
-                    latency_ms = latency.as_millis(),
+                    latency_ms = latency.as_millis() as u64,
                     "HTTP request completed"
                 );
             },
